refactor: organize classes into packages
