+++ src/main/java/com/example/sla/repository/SlaActionLogRepository.java
@@ -1,6 +1,8 @@
-package com.example.sla;
+package com.example.sla.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.sla.domain.SlaActionLog;
 
 public interface SlaActionLogRepository extends JpaRepository<SlaActionLog, Long> {
     boolean existsByTaskIdAndAction(String taskId, String action);
