+++ src/main/java/com/example/sla/repository/SlaInstanceRepository.java
@@ -1,7 +1,9 @@
-package com.example.sla;
+package com.example.sla.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.sla.domain.SlaInstance;
 
 public interface SlaInstanceRepository extends JpaRepository<SlaInstance, String> {
     List<SlaInstance> findByStatus(String status);
